+++ app/containers/ProfilePage/definitions.ts
@@ -1,0 +1,36 @@
+import icons from "@/app/components/icons";
+
+export const SettingsDefinition = [
+  {
+    icon: icons.calendar,
+    name: "My Bookings",
+  },
+  {
+    icon: icons.wallet,
+    name: "Payments",
+  },
+  {
+    icon: icons.person,
+    name: "Profile",
+  },
+  {
+    icon: icons.bell,
+    name: "Notifications",
+  },
+  {
+    icon: icons.shield,
+    name: "Security",
+  },
+  {
+    icon: icons.language,
+    name: "Language",
+  },
+  {
+    icon: icons.info,
+    name: "Help Center",
+  },
+  {
+    icon: icons.people,
+    name: "Invite Friends",
+  },
+];
